refactor: rename function in ConversationService
