+++ src/ChitChat.WebAPI/Controllers/UserController.cs
@@ -1,6 +1,7 @@
-﻿using ChitChat.Application.Models;
+using ChitChat.Application.Models;
 using ChitChat.Application.Models.Dtos.User;
 using ChitChat.Application.Services.Interface;
+
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
